requirements.txt
# Moltbook Analysis Pipeline Requirements

# Core
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0  # For parquet support

# API & HTTP
requests>=2.31.0
python-dotenv>=1.0.0  # For .env file support

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0  # For model serialization

# Topic Modeling (optional but recommended)
bertopic>=0.15.0
hdbscan>=0.8.33

# Visualization (optional but recommended)
matplotlib>=3.7.0
seaborn>=0.12.0

# Scientific computing
scipy>=1.11.0

# Development
pytest>=7.0.0
black>=23.0.0
